+++ src/ipyaladin/aladin.py
@@ -11,10 +11,10 @@ from typing import ClassVar, Union, Final, Optional
 import warnings
 
 import anywidget
+from astropy.coordinates import SkyCoord, Angle
 from astropy.table.table import QTable
 from astropy.table import Table
-from astropy.coordinates import SkyCoord, Angle
-import traitlets
+from astropy.wcs import WCS
 
 try:
     from regions import (
@@ -34,6 +34,7 @@ except ImportError:
     RectangleSkyRegion = None
     Region = None
     Regions = None
+import traitlets
 from traitlets import (
     Float,
     Int,
@@ -121,6 +122,9 @@ class Aladin(anywidget.AnyWidget):
     grid_color = Unicode("rgb(178, 50, 178)").tag(sync=True, init_option=True)
     grid_opacity = Float(0.5).tag(sync=True, init_option=True)
     grid_options = traitlets.Dict().tag(sync=True, init_option=True)
+
+    # Synchronized traitlets
+    _wcs = traitlets.Dict().tag(sync=True)
 
     # content of the last click
     clicked_object = traitlets.Dict().tag(sync=True)
@@ -219,6 +223,33 @@ class Aladin(anywidget.AnyWidget):
                 "dec": target.icrs.dec.deg,
             }
         )
+
+    @property
+    def wcs(self) -> WCS:
+        """The WCS of the Aladin Lite widget.
+
+        You need to call synchronize_wcs in a previous
+        cell before accessing this attribute.
+
+        Returns
+        -------
+        WCS
+            An `~astropy.wcs.WCS` object representing the WCS of the widget.
+
+        """
+        if self._wcs == {}:
+            raise ValueError(
+                "You need to call synchronize_wcs before accessing the wcs attribute. "
+                "The WCS attribute need to be accessed from the next cell."
+            )
+        if "RADECSYS" in self._wcs:  # RADECSYS keyword is deprecated for astropy.WCS
+            self._wcs["RADESYS"] = self._wcs.pop("RADECSYS")
+        return WCS(self._wcs)
+
+    def synchronize_wcs(self) -> None:
+        """Synchronize the WCS of the Aladin Lite widget with the given WCS."""
+        self._wcs = {}
+        self.send({"event_name": "synchronize_wcs"})
 
     def add_catalog_from_URL(
         self, votable_URL: str, votable_options: Optional[dict] = None